chatbot/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.5
python-dotenv==1.1.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-ollama==0.3.3
langchain-huggingface==0.3.0
huggingface-hub==0.30.2
sentence-transformers==4.1.0
faiss-cpu==1.11.0
python-docx==1.1.2
orjson==3.10.18
tqdm==4.67.1
numpy
#numpy == 2.2.5  (I change first to see if it works lol -kx)
regex==2024.11.6
uvicorn
rapidfuzz